requirements.txt
Flask==2.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
ytmusicapi==1.2.0
